Update vacuum.xiaomi_miio.markdown
This section of the vacuum page is linked to extensively by other pages explaining how to get the Xiaomi token code. It needs to be made clearer that each code is unique and this will only gives the hub code. I had to use the iOS method to obtain unique codes for my Vac and Wifi Plugs.